requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
google-adk
yfinance
pandas
pydantic
python-multipart